refactor(swarm-components): charcount as standalone file in shared dir
